fix some bugs of one to one mapping change some annotations for entity id

+++ java/javaeeHomeworkEJB/src/service/ConsumeService.java
@@ -20,5 +20,5 @@ public interface ConsumeService {
     boolean payMoney(int uid, double money);
 
     List<Consume> userBalanceList();
-    int countUserBalanceList();
+    long countUserBalanceList();
 }
