+++ java/javaeeHomeworkEJB/src/service/MessageService.java
@@ -16,5 +16,5 @@ public interface MessageService {
     void addMsg(String msg, int toUser);
     void deleteMsg(int mid);
     ArrayList<Message> userMsg(int uid, int startIndex, int pageSize);
-    int countUserMsg(int uid);
+    long countUserMsg(int uid);
 }
